Changes for User Story "LAS Links - Customer Received Random Student IDs (Defect#77134)"

+++ dep-orgmgmt-ctrl/src/com/ctb/util/UploadStudent.java
@@ -3675,8 +3675,8 @@ System.out.println("studentIdList.contains(strCell.trim()) : "+studentIdList.con
 
 				if ( !datFormat ) {
 
-					cellValue = String.valueOf((new Double (cell.getNumericCellValue())).
-							intValue());
+					cellValue = String.valueOf((new Double (cell.getNumericCellValue())).longValue());
+							
 				}
 
 				// date formatted is true, then format it to mm/dd/yyyy
